1.From Lection 2.Add Calculator

+++ src/main/java/pro/sky/skypromavenspring/TextController.java
@@ -11,33 +11,39 @@ public class TextController {
     public TextController(TextServiceInterface textService) {
         this.textService = textService;
     }
+
     @GetMapping
     public String hello() {
         return textService.hello();
     }
 
-    @GetMapping(path= "/hello")
-    public String answerHello(@RequestParam ("name") String userName) {
+    @GetMapping(path = "/hello")
+    public String answerHello(@RequestParam("name") String userName) {
         return textService.answerHello(userName);
     }
-    @GetMapping(path= "/calculator")
+
+    @GetMapping(path = "/calculator")
     public String calculatorPage() {
         return textService.calculatorPage();
     }
-    @GetMapping(path= "/calculator/plus")
-    public double calculatorPageCalcPlus(@RequestParam ("num1") double valueOne, @RequestParam ("num2") double valueTwo) {
+
+    @GetMapping(path = "/calculator/plus")
+    public int calculatorPageCalcPlus(@RequestParam("num1") int valueOne, @RequestParam("num2") int valueTwo) {
         return textService.calculatorPageCalcPlus(valueOne, valueTwo);
     }
-    @GetMapping(path= "/calculator/minus")
-    public double calculatorPageCalcMinus(@RequestParam ("num1") double valueOne, @RequestParam ("num2") double valueTwo) {
+
+    @GetMapping(path = "/calculator/minus")
+    public int calculatorPageCalcMinus(@RequestParam("num1") int valueOne, @RequestParam("num2") int valueTwo) {
         return textService.calculatorPageCalcMinus(valueOne, valueTwo);
     }
-    @GetMapping(path= "/calculator/multiply")
-    public double calculatorPageCalcMul(@RequestParam ("num1") double valueOne, @RequestParam ("num2") double valueTwo) {
+
+    @GetMapping(path = "/calculator/multiply")
+    public int calculatorPageCalcMul(@RequestParam("num1") int valueOne, @RequestParam("num2") int valueTwo) {
         return textService.calculatorPageCalcMul(valueOne, valueTwo);
     }
-    @GetMapping(path= "/calculator/divide")
-    public double calculatorPageCalcDiv(@RequestParam ("num1") double valueOne, @RequestParam ("num2") double valueTwo) {
+
+    @GetMapping(path = "/calculator/divide")
+    public double calculatorPageCalcDiv(@RequestParam("num1") int valueOne, @RequestParam("num2") int valueTwo) {
         return textService.calculatorPageCalcDiv(valueOne, valueTwo);
     }
